refactor(registration): 검증 부분 분리하여 파일 생성
